bfx: Drop superfluous carriage return
When timer displays the remaining time on stdout,
it uses a carriage return (\r) to repeatedly "update" a single line
rather than constantly appending output to the screen every second.

This carriage return was previously attached to a variable
that was also used to set the tmux session name and the window title,
and that superfluous carriage return was visible in some cases.

This commit isolates the carriage return (\r character)
to the logic that prints to stdout.