ftr: Use colons in countdown formatting
This commit changes the format of countdown text:

    00h04m20s  # old
    04:20      # new

Why? Because it's prettier this way.

=== Caveat: tmux

Unfortunately, tmux doesn't support colons (or dots) in window names.
I haven't fully investigated the reason for this,
but it appears likely that these characters have special meaning
in tmux command mechanisms.

As a result, the old formatting prevails in tmux window names.