bfx: Fix bash syntax error for delay (-d) option
When passing a value for the `-d` option,
timer gives the following error output:

    $ timer -d 1 0
    00:01/opt/homebrew/bin/timer: line 58: [: -eq: unary operator expected

This occurs because the `ring()` function expects a `set_count` var,
which is only ever set via the `run_n_sets()` function.
That function is invoked for everything except the initial delay timer;
this commit resolves the error by adding an extra `if` clause
for the case where `set_count` is empty.

Fixes #6.